Shop improvements
now able to load in items into the shop and should be selectable now. Current button format is not working so going to try to remake it so it isn't as buggy. Up next is to move on to making a basic currency system for the game before starting save states.