fixed Shop and got functionality
Was able to get functionality to work for selecting an item. Item background now enables when an Item is selected and is disabled when another item is selected. scriptable item information in ShopItem list can now be passed over to different facets of the description. Some information needs to be added to scriptable items but can all be added pretty quickly. Last thing that needs to be added to the shop on a implementation standpoint is the action of buying and adding items, which also comes with implementing an economy system.