Added Cost and Buy button
There is now a cost section and buy button. It brings up a panel that prompts the player to whether they want to purchase the Item. Doesn't actually do anything but changes the amount of money in the wallet which is in the top right. The shop doesn't take any info from items selected. so the next update will be part of getting that functionality working and then implementing on to the level scenes.