Save Update
Updated Save and Load system so that It doesn't just automatically save when exiting the scene. Also fixed a lot of the different bugs where the start button would load the last saved data even though it should start a new game. Also took out the exitlevelbutton and made it so that if the escape button is pressed, brings up a new panel where it prompts the player to save or not save their inforation. Now all that is needed is to implement this into the battle scene transitions so that the player doesn't restart their position. Also the introduction of saving other information other than health and transform.