First try at new data persistence with battle scene
Tried making it so that data can transfer between the level scene and the player scene. I was able to get it to where it can transfer the data from the level scene to the battle scene, although it causes a lot of bugs. Sometimes data transfers to the level scene back from the battle scene but sometimes it doesn't. It also assigns the player health to the health of the bat. Then sometimes it doesn't transfer back at all. Still working out a lot of the bugs. Although initial idea is working though.