Hit Effects for Combat Scene
Added setup for adding hit animations for attacks, positioning may need some work. Also, for the unit death issue, The issue involves the having the tag, "DeadUnit". Adding the tag back in will fix it, but GitHub won't commit it so that's up to everyone to add in I guess.