Duplicate Enemies, Tried Reward System.
Duplication is solved forever.
Enemy spawners have manually assigned unique IDs to check for duplication.
The enemy spawner will permanently deactivate upon entering battle, need some check on if the player ran or not to re-enable it.

Enemy encounter prefab, (which holds what units to spawn when entering the spawner), holds the variable and method for the enemy encounter's reward. The player is rewarded in the battle scene by increasing their currency value in the player's UnitStats script. Need some help on why UnitStats resets though.